% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apextra.r
\docType{package}
\name{apextra}
\alias{apextra}
\alias{apextra-package}
\title{R package with additional functions to supplement the existing package "ape"}
\description{
The package \pkg{apextra} provides additional functions to supplement the existing package "ape"
}
\author{
Chris Field <fieldc@ethz.ch">
}
