% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.r
\name{as.ultrametric}
\alias{as.ultrametric}
\title{Function to extend the terminal branches of a phylo object so that the resulting tree is ultrametric}
\usage{
as.ultrametric(phylo)
}
\arguments{
\item{phylo}{An object of class "phylo" that is rooted.}
}
\value{
An object of class "phylo", rooted and ultrametric
}
\description{
Function to extend the terminal branches of a phylo object so that the resulting tree is ultrametric
}
\details{
The final root-to-tip distance of each tip will be that of the furthest from the root.
}
\examples{
None
}
\author{
Chris Field <fieldc@ethz.ch>
}
\keyword{None}
