% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_functions.r
\name{draw.phylo}
\alias{draw.phylo}
\title{Function to draw a phylogram on an existing plot in an arbitrary position with arbitrary scale.}
\usage{
draw.phylo(x1, y1, x2, y2, phylo, direction = "r", mirror = F,
  show.tip.label = FALSE, label.offset = 0, tip.color = "black",
  align.tip.label = FALSE)
}
\arguments{
\item{x1, y1, x2, y2}{Coordinates of the rectangle that will contain the phylogram.}

\item{phylo}{An object of class "phylo".}

\item{direction}{A character string specifying the direction in which the phylogram will be drawn; it must be one of "r"ightwards (default), "l"eftwards, "u"pwards or "d"ownwards.}

\item{mirror}{A logical indicating whether the phylogeny should be plotted in reverse order, i.e.: mirrored.}

\item{show.tip.label}{A logical indicating whether tip labels should be drawn; defaults to FALSE. The labels will be placed outside of the coordinate rectangle provided.}

\item{label.offset}{A numeric giving the distance between the tips and their corresponding labels, defaults to 0.}

\item{tip.color}{A single color or vector of colors to be used for the tip labels, defaults to black.}

\item{align.tip.label}{A logical indicating whether tip labels should be aligned with the furthest tip rather than each tip individually, defaults to FALSE.}
}
\value{
None
}
\description{
Function to draw a phylogram on an existing plot in an arbitrary position with arbitrary scale.
}
\details{
The phylogram is laid out such that the outermost branches lie on the edges of the coordinate rectangle, with the root and furthest tip on the other edges.
}
\examples{
None
}
\author{
Chris Field <fieldc@ethz.ch>
}
\keyword{None}
