% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_functions.r
\name{higherNodes}
\alias{higherNodes}
\title{Function to find the nodes one level up from a provided set of nodes}
\usage{
higherNodes(phylo, nodes)
}
\arguments{
\item{phylo}{An object of class "phylo".}

\item{nodes}{A numeric vector indicating the nodes to start from. Tips are also nodes, numbered 1:N.}
}
\value{
An object of class "phylo".
}
\description{
Function to find the nodes one level up from a provided set of nodes
}
\examples{
None
}
\author{
Chris Field <fieldc@ethz.ch>
}
\keyword{None}
