% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_functions.r
\name{nearestNeighbours}
\alias{nearestNeighbours}
\title{Function to find the nearest neighbour(s) to a provided set of tips}
\usage{
nearestNeighbours(phylo, tips, n = 1)
}
\arguments{
\item{phylo}{An object of class "phylo".}

\item{tips}{A numeric vector indicating the tips to start from.}

\item{n}{A numeric indicating the number of nearest neighbours to return.}
}
\value{
An object of class "phylo".
}
\description{
Function to find the nearest neighbour(s) to a provided set of tips
}
\examples{
None
}
\author{
Chris Field <fieldc@ethz.ch>
}
\keyword{None}
