% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_functions.r
\name{reorderTips}
\alias{reorderTips}
\title{Function to reorder the tip labels of a tree to match the order in which they are plotted, which can change when a tree is rerooted}
\usage{
reorderTips(phylo)
}
\arguments{
\item{phylo}{An object of class "phylo".}
}
\value{
An object of class "phylo"
}
\description{
Function to reorder the tip labels of a tree to match the order in which they are plotted, which can change when a tree is rerooted
}
\details{
First the internal labels in the edges of the tree are relabelled to appear in ascending order, then the tip labels are reordered to match the correct internal labels.
}
\examples{
None
}
\author{
Chris Field <fieldc@ethz.ch>
}
\keyword{None}
