% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_functions.r
\name{rev.phylo}
\alias{rev.phylo}
\title{Function to reverse the structure of a phylogeny for plotting purposes.}
\usage{
\method{rev}{phylo}(phylo)
}
\arguments{
\item{phylo}{An object of class "phylo".}
}
\value{
An object of class "phylo".
}
\description{
Function to reverse the structure of a phylogeny for plotting purposes.
}
\examples{
None
}
\author{
Chris Field <fieldc@ethz.ch>
}
\keyword{None}
