% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_functions.r
\name{tipOrder}
\alias{tipOrder}
\title{Function to determine the 'real' order of tips in a phylo object, which can change when a tree is rerooted}
\usage{
tipOrder(phylo)
}
\arguments{
\item{phylo}{An object of class "phylo".}
}
\value{
A vector with the tip numbers 1 to N in the order they are in the tree.
}
\description{
Function to determine the 'real' order of tips in a phylo object, which can change when a tree is rerooted
}
\details{
None
}
\examples{
None
}
\author{
Chris Field <fieldc@ethz.ch>
}
\keyword{None}
