% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_functions.r
\name{tipsBelow}
\alias{tipsBelow}
\title{Function to find the tips below a provided set of nodes}
\usage{
tipsBelow(phylo, nodes)
}
\arguments{
\item{phylo}{An object of class "phylo".}

\item{nodes}{A numeric vector indicating the nodes to start from.}
}
\value{
An object of class "phylo".
}
\description{
Function to find the tips below a provided set of nodes
}
\examples{
None
}
\author{
Chris Field <fieldc@ethz.ch>
}
\keyword{None}
